♻️ factor out socket devtools
